+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - #251 - Shutdown upon non responding (unreachable) cluster is not possible
 - #258 - Investigate lowering requirements on activesupport
 - Replaced some of the activesupport parts with dry-inflector
+- #246 - Alias consumer#mark_as_consumed on controller
 
 ## 1.1.2
 - #256 - Default kafka.seed_brokers configuration is created in invalid format